:truck: Rename files from About to Home
